feat: Improve the interface of the control center
Improve the interface of the control center

Log:
Task: https://pms.uniontech.com/task-view-361721.html

+++ src/dde-control-center/controlcenterdbusadaptor.cpp
@@ -119,7 +119,7 @@ QString ControlCenterDBusAdaptor::GetAllModule()
         QJsonObject obj;
         obj.insert("url", urlInfo.second.at(0));
         obj.insert("displayName", urlInfo.second.at(1));
-        obj.insert("weight", urlInfo.first->weight());
+        obj.insert("weight", (int)(urlInfo.first->weight()));
         arr.append(obj);
         const QList<DccObject *> &children = urlInfo.first->getChildren();
         for (auto it = children.crbegin(); it != children.crend(); ++it)
